refactor to support single-binary packaging
node and electron are messy.  This restructures the gpc
simulator to make it easier to use a packaging tool to
generate a single binary that has no (/minimal) dependencies.

Since gpc-gui already uses electron, we use electron-builder
to generate a linux .AppImage or Mac .App

Instead of having separate cli tools for batch, debugger,
debugger gui and fcm dump, there's now a single 'gpc' multicommand:

❯ ./GPC.sh

  dist/gpc.js  614.2kb

⚡ Done in 690ms
Usage: gpc [options] [command]

AP-101 GPC Simulator

Options:
  -V, --version                   output the version number
  -h, --help                      display help for command

Commands:
  run [options] <fcm-file>        Run an AP-101 program in batch mode
  debug|dbg [options] <fcm-file>  Interactive AP-101 debugger
  gui [options] [fcm-file]        Electron GUI debugger
  dump [options] <fcm-file>       FCM dump report — symbol table and disassembly
  disasm [options] <fcm-file>     Disassemble an FCM memory image
  help [command]                  display help for command

If you use an AppImage or App, it should work the same.

To get this all to work, I refactored the separate commands to
share share more code.  I threw them together super fast initially,
so this is a good cleanup in any case.  While doing this, I rearranged
how functions are split up conceptually so:

- pull all the debugging/gui stuff out of AP101 and make it closer to
  a pure representation of the hardware.
    - NOTE: it's still a mix of AP-101-B and AP-101-S behavior.
      When I get a chance to properly finish the AP-101-S specific
      functions I'll split it into AP101B and AP101S to isolate the
      differences.
- AGEHarness sits above AP101 and provides the basic debugging &
  development functions.  The HalUCP ('User Control Program') that
  implements I/O and error handling in the dev environment attaches
  here, along with loading memory, symbols, breakpoints, etc.
  AGE = "Aerospace/Ground Equipment", the term IBM used to refer
  to develoment hardware that would connect to the AP-101's in
  ground environments.

- GUIHarness adds stuff you need for running the interactive debugger

- the top level run_ files become cmd_ files used by the 'gpc' top
  level.

Build also changes:

  - setup build system to package for linux and mac
  - add a github action to build linux AppImage
