+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 dist/
+release/
 *.log
 .DS_Store
 Thumbs.db
