--- config/gpc-layout.json
@@ -1,7 +1,0 @@
-{
-  "watchWidth": 130,
-  "regWidth": 380,
-  "memHeight": 235,
-  "sectionWidth": 387,
-  "terminalHeight": 335
-}
