--- config/gpcDev.json
@@ -1,21 +1,0 @@
-{
-  "start": [
-    "gpc1"
-  ],
-  "startSharedWindow": false,
-  "lrus": {
-    "gpc1": {
-      "window": {
-        "width": 1800,
-        "height": 1333,
-        "x": 50,
-        "y": 10,
-        "backgroundColor": "#111111"
-      },
-      "config": {
-        "fcmFile": "gpc/gen/SIMPLE.fcm",
-        "symbolsFile": "gpc/gen/SIMPLE.fcm.LIST"
-      }
-    }
-  }
-}
