--- config/mdu.yaml
@@ -1,86 +1,0 @@
-DISABLED_start:
-  - idp1
-  - idp2
-  - idp3
-  - idp4
-  - cdr1
-start:
-  - idp1
-  - cdr1
-startSharedWindow: true
-lrus:
-  shared:
-    window:
-      width: 600
-      height: 800
-      x: 10
-      y: 10
-      backgroundColor: "#444444"
-      fullscreen: false
-  ccms1:
-    lru: CCMSConsole
-    module: lps/ccms
-    window:
-      widght: 1280
-      height: 510
-      x: 300
-      y: 50
-      backgroundColor: rgb(191,168,158)
-      fullscreen: false
-    config:
-      lru: CCMS1
-  gpc1:
-    lru: GPC
-    module: gpc/ap101b
-    window:
-      width: 600
-      height: 800
-      x: 620
-      y: 10
-      backgroundColor: "#444444"
-      fullscreen: false
-    config:
-      lru: GPC1
-  idp1:
-    shared: true
-    lru: IDP
-    module: meds/idp
-    config:
-      lru: IDP1
-  idp2:
-    shared: true
-    lru: IDP
-    module: meds/idp
-    config:
-      lru: IDP2
-  idp3:
-    shared: true
-    lru: IDP
-    module: meds/idp
-    config:
-      lru: IDP3
-  idp4:
-    shared: true
-    lru: IDP
-    module: meds/idp
-    config:
-      lru: IDP4
-  cdr1:
-    lru: MDU
-    window:
-      width: 720
-      height: 720
-      x: 50
-      y: 50
-      fullscreen: false
-      html: meds/index.html
-    module: meds/mdu
-    enableTouchbar: true
-    supersample: 4
-    widthThinLine: 0.2
-    widthThickLine: 0.15
-    config:
-      lru: CDR2
-    init:
-      menu: FLT_INST
-      display: AE_PFD
