--- config/mduPFD.yaml
@@ -1,4 +1,0 @@
-start:
-  - idp1
-  - cdr1
-startSharedWindow: true
