--- config/mdu_1152.yaml
@@ -1,19 +1,0 @@
-- start: idp
-  idp:
-    config:
-      lru: IDP1
-- start: mdu
-  mdu:
-    width: 1152
-    height: 1152
-    x: 100
-    y: 100
-    fullscreen: false
-    supersample: 4
-    widthThinLine: 0.2
-    widthThickLine: 0.15
-    config:
-      lru: PLT2
-    init:
-      menu: DISCONNECTED
-      display: AUTONOMOUS
