--- config/mtu.yaml
@@ -1,3 +1,0 @@
-start:
-  - MDM_FF1
-  - MTU1
