+++ simRunner/renderer/startup.civet
@@ -1,25 +1,14 @@
 "civet coffeeCompat"
 
-import * as gpc from './../../gpc/run_gui.coffee'
+import * as gpc from './../../gpc/gui.coffee'
 
 ipcRenderer := window.ipcRenderer
 reactClient := require 'react-dom/client';
 
-ipcRenderer.on "set-config", async (e,CONFIG) =>
-  console.log("CONFIG --> ", CONFIG)
-  for lruName in CONFIG.thisStart
-    console.log("NAME: ", lruName)
-    lruConf .= CONFIG.lrus[lruName]
-    console.log("START LRU", lruConf)
-    lruConf.NSTS_TOP = CONFIG.NSTS_TOP
-    lruMap .= {
-      'gpc/ap101': gpc
-    }
-    console.log(lruMap[lruConf.module])
-    lru .= lruMap[lruConf.module].default.start(lruConf)
-    if lru.initWindow?
-      container .= document.getElementById("root")
-      await reactClient.createRoot(container!).render(await lru.initWindow())
-    console.log("LRU",lru)
-    window.lru = lru
-    lru.start()
+ipcRenderer.on "set-config", async (e, CONFIG) =>
+  debugGui := gpc.default.start(CONFIG)
+  if debugGui.initWindow?
+    container := document.getElementById("root")
+    await reactClient.createRoot(container!).render(await debugGui.initWindow())
+  window.lru = debugGui
+  debugGui.start()
